Fixed bug when opening app after first time
